Fix for trailing '.' in variable assignment
Note that `substr` in Terraform does not use -1 like other languages.

This code is also compatible with hashicorp/aws v3. As additionally, there was a
very confusing scenario whereby later versions of the hashicorp/aws provider
return `aws_route53_zone.name` _without_ a trailing '.'.  This meant that
because we do not have version pinning of providers, a local Terraform run would
use a newer version of providers than those running in prod which meant that
local runs would work!